+++ src/classes/org/tritonus/share/midi/MidiUtils.java
@@ -44,7 +44,7 @@ public class MidiUtils
 {
 	public static int getUnsignedInteger(byte b)
 	{
-		return (b < 0) ? (int) b + 256 : b;
+		return (b < 0) ? b + 256 : b;
 	}
